added abstract class Function
